refactor: use KokkosCore builtin reducers instead of custom ones
Directly use KokkosCore's builtin reducers instead of re-declaring ones
in KokkosComm.

Only the `Average` reducer is declared in KokkosComm since it isn't
provided by Core and is an available option in NCCL. It is merely a
"marker" struct though, since we do not need to implement anything
ourselves, it only serves to describe to NCCL what reduction operator we
want to use.