ci: run NCCL unit tests on SNL CI
- NCCL is built from source, packaged into an OS-agnostic tarball &
installed from that tarball. For now, I am unsure if this will actually
work in the container.
- Kokkos version is explicitly set to 4.7.01 (equivalent to master at
the time of this commit). In the future, we hope to have a container
that installs the latest versions of Kokkos/Open MPI/MPICH/NCCL and have
it cached by GHA. Submitted CI jobs will just pull down this container
image with all dependencies set up, and only build KokkosComm.
- For now, KokkosComm's MPI and NCCL backends are
configured/built/tested one after the other. This may be improved later
on, as both are independent and could theoretically run concurrently.